Move Name generation into its own class
There was no shared logic at all with JSGen. This gives us two
advantages:

- Keep private things more private,
- Keep name caches alive for longer.
